toolkit/content/tests/chrome/window_subframe_origin.xul

<?xml version="1.0"?>
<!-- ***** BEGIN LICENSE BLOCK *****
   - Version: MPL 1.1/GPL 2.0/LGPL 2.1
   -
   - The contents of this file are subject to the Mozilla Public License Version
   - 1.1 (the "License"); you may not use this file except in compliance with
   - the License. You may obtain a copy of the License at
   - http://www.mozilla.org/MPL/
   -
   - Software distributed under the License is distributed on an "AS IS" basis,
   - WITHOUT WARRANTY OF ANY KIND, either express or implied. See the License
   - for the specific language governing rights and limitations under the
   - License.
   -
   - The Original Code is Findbar Test code
   -
   - The Initial Developer of the Original Code is
   - Mozilla Foundation.
   - Portions created by the Initial Developer are Copyright (C) 2010
   - the Initial Developer. All Rights Reserved.
   -
   - Contributor(s):
   -
   - Alternatively, the contents of this file may be used under the terms of
   - either the GNU General Public License Version 2 or later (the "GPL"), or
   - the GNU Lesser General Public License Version 2.1 or later (the "LGPL"),
   - in which case the provisions of the GPL or the LGPL are applicable instead
   - of those above. If you wish to allow use of your version of this file only
   - under the terms of either the GPL or the LGPL, and not to allow others to
   - use your version of this file under the terms of the MPL, indicate your
   - decision by deleting the provisions above and replace them with the notice
   - and other provisions required by the GPL or the LGPL. If you do not delete
   - the provisions above, a recipient may use your version of this file under
   - the terms of any one of the MPL, the GPL or the LGPL.
   -
   - ***** END LICENSE BLOCK ***** -->
<?xml-stylesheet href="chrome://global/skin" type="text/css"?>

<window id="window" title="Subframe Origin Tests"
        xmlns="http://www.mozilla.org/keymaster/gatekeeper/there.is.only.xul">
<script type="application/javascript"
        src="chrome://mochikit/content/tests/SimpleTest/EventUtils.js"/>

<iframe
  style="margin-left:20px; margin-top:20px; min-height:300px; max-width:300px; max-height:300px; border:solid 1px black;"
  src="frame_subframe_origin_subframe1.xul"></iframe>
<caption id="parentcap" label=""/>

<script>

// Fire a mouse move event aimed at this window, and check to be
// sure the client coords translate from widget to the dom correctly.

function runTests()
{
  synthesizeMouse(document.getElementById("window"), 1, 2, { type: "mousemove" });
}

window.opener.wrappedJSObject.SimpleTest.waitForFocus(runTests, window);

function mouseMove(e) {
  var element = e.target;
  var el = document.getElementById("parentcap");
  el.label = "client: (" + e.clientX + "," + e.clientY + ")";
  window.opener.wrappedJSObject.SimpleTest.is(e.clientX, 1, "mouse event clientX");
  window.opener.wrappedJSObject.SimpleTest.is(e.clientY, 2, "mouse event clientY");
  // fire the next test on the sub frame
  frames[0].runTests();
}

window.addEventListener("mousemove",mouseMove, false);

</script>
</window>
